% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_globals.R
\name{get_globalCapabilities}
\alias{get_globalCapabilities}
\title{Create master list of variable lookups for "capabilities" (output variables for graphing)}
\usage{
get_globalCapabilities()
}
\description{
Create master list of variable lookups for "capabilities" (output variables for graphing)
}
