% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_globals.R
\name{get_globalParameters}
\alias{get_globalParameters}
\title{Create master list of parameter lookup strings}
\usage{
get_globalParameters()
}
\description{
Create master list of parameter lookup strings
}
