% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_globals.R
\name{get_globalVars}
\alias{get_globalVars}
\title{Global vars for misc items such as the run date end year (2100)}
\usage{
get_globalVars()
}
\description{
Global vars for misc items such as the run date end year (2100)
}
