% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{constants}
\alias{constants}
\alias{globalVars}
\alias{globalScenarios}
\alias{globalParameters}
\alias{globalCapabilities}
\title{Global constants for the Hector-Shiny UI}
\format{An object of class \code{numeric} of length 1.}
\usage{
globalVars

globalScenarios

globalParameters

globalCapabilities
}
\description{
Global constants for the Hector-Shiny UI
}
\details{
\code{globalVars}: Miscellaneous global variables

\code{globalScenarios} Scenario input files

\code{globalParameters}: Capability strings for Hector parameters

\code{globalCapabilities} Capability strings for Hector variables
}
\keyword{datasets}
