% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{constants}
\alias{constants}
\alias{globalVars}
\alias{globalColorScales}
\alias{globalScenarios}
\alias{globalParameters}
\alias{globalParamsDefault}
\alias{globalParamsCanESM2}
\alias{globalParamsCESM1BGC}
\alias{globalParamsGFDLESM2G}
\alias{globalParamsMIROCESM}
\alias{globalParamsMPIESMLR}
\alias{globalParamsMRIESM1}
\alias{globalCapabilities}
\title{Global constants for the Hector-Shiny UI}
\format{An object of class \code{numeric} of length 2.}
\usage{
globalVars

globalColorScales

globalScenarios

globalParameters

globalParamsDefault

globalParamsCanESM2

globalParamsCESM1BGC

globalParamsGFDLESM2G

globalParamsMIROCESM

globalParamsMPIESMLR

globalParamsMRIESM1

globalCapabilities
}
\description{
Global constants for the Hector-Shiny UI
}
\details{
\code{globalVars}: Miscellaneous global variables

\code{globalColorScales}: Scale colors

\code{globalScenarios} Scenario input file names and paths

\code{globalParameters}: Capability strings (internal name lookup/mapping) for Hector parameters - (should match UI component ids)

\code{globalParamsDefault}: Default Parameter Sets

\code{globalParamsCanESM2}: CanESM2 Parameters for model emulation

\code{globalParamsCESM1BGC}: CESM1-BGC Parameters for model emulation

\code{globalParamsGFDLESM2G}: GFDL-ESM2G Parameters for model emulation

\code{globalParamsMIROCESM}: MIROC-ESM Parameters for model emulation

\code{globalParamsMPIESM-LR}: MPI-ESM-LR Parameters for model emulation

\code{globalParamsMRIESM1}: MRI-ESM1 Parameters for model emulation

\code{globalCapabilities} Capability string (internal name lookup/mapping) for Hector output variables
}
\keyword{datasets}
