% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{constants}
\alias{constants}
\alias{globalVars}
\alias{globalScenarios}
\alias{globalParameters}
\alias{globalParamsDefault}
\alias{globalParamsCanESM2}
\alias{globalParamsCESM1BGC}
\alias{globalParamsGFDLESM2G}
\alias{globalParamsMIROCESM}
\alias{globalParamsMPIESMLR}
\alias{globalParamsMRIESM1}
\alias{globalCapabilities}
\title{Global constants for the Hector-Shiny UI}
\format{An object of class \code{numeric} of length 2.}
\usage{
globalVars

globalScenarios

globalParameters

globalParamsDefault

globalParamsCanESM2

globalParamsCESM1BGC

globalParamsGFDLESM2G

globalParamsMIROCESM

globalParamsMPIESMLR

globalParamsMRIESM1

globalCapabilities
}
\description{
Global constants for the Hector-Shiny UI
}
\details{
\code{globalVars}: Miscellaneous global variables

\code{globalScenarios} Scenario input files

\code{globalParameters}: Capability strings for Hector parameters

\code{globalParamsDefault}: Default Parameter Sets

\code{globalParamsCanESM2}: Default Parameter Sets

\code{globalParamsCESM1BGC}: Default Parameter Sets

\code{globalParamsGFDLESM2G}: Default Parameter Sets

\code{globalParamsMIROCESM}: Default Parameter Sets

\code{globalParamsMPIESM-LR}: Default Parameter Sets

\code{globalParamsMRIESM1}: Default Parameter Sets

\code{globalCapabilities} Capability strings for Hector variables
}
\keyword{datasets}
